revert change to int16_t for cities, needed for dist matrix

+++ tsp/greedy.cpp
@@ -89,8 +89,7 @@ int main(int argc, char *argv[]) {
 
   srandom(seed);
 
-  RR_greedy<random_access_list<int16_t>,
-            std::vector<int16_t>>(argv[optind], seed);
+  RR_greedy<random_access_list<int>, std::vector<int>>(argv[optind], seed);
 
   return EXIT_SUCCESS;
 }
